+++ node/module.mjs
@@ -48,7 +48,6 @@ export const validateString = (json, options) => v.default.validateString(json, 
  @typedef {Object} ValidationOptions
  @property {string} uri - Absolute or relative asset URI that will be copied to validation report.
  @property {ExternalResourceFunction} externalResourceFunction - Function for loading external resources. If omitted, external resources are not validated.
- @property {boolean} validateAccessorData - Set to `false` to skip reading of accessor data.
  @property {boolean} writeTimestamp - Set to `false` to omit timestamp from the validation report.
  @property {number} maxIssues - Max number of reported issues. Use `0` for unlimited output.
  @property {string[]} ignoredIssues - Array of ignored issue codes.